+++ crates/tests/src/lib.rs
@@ -23,4 +23,8 @@ mod timeline;
 mod host_server;
 
 #[cfg(test)]
+#[cfg(feature = "p2p")]
+mod p2p;
+
+#[cfg(test)]
 mod multi_server;